feat(issue180): Architecture refactors to prepare the project to migrate to fastAPI.
