--- .codclimate.yml
@@ -1,5 +1,0 @@
-languages:
-  Python: true
-plugins:
-  pep8:
-    enabled: true
