--- .covaragerc
@@ -1,2 +1,0 @@
-[run]
-source=cotacoes
