--- .github/workflows/main.yml
@@ -1,31 +1,0 @@
-name: CI
-on:
-  push:
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  ci:
-    strategy:
-      fail-fast: true
-      matrix:
-        python-version: ["3.10"]
-        poetry-version: ["1.2.1"]
-        os: [ubuntu-latest]
-    runs-on: ${{ matrix.os }}
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Run image
-        uses: abatilo/actions-poetry@v2
-        with:
-          poetry-version: ${{ matrix.poetry-version }}
-      - name: Install dependencies
-        run: poetry install
-      - name: Lint
-        run: poetry run flake8
-      - name: Test
-        run: poetry run pytest --cov='.'
