+++ .gitignore
@@ -1,21 +1,9 @@
-# Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
 *$py.class
-tmp_cotar_bceuropeu.py
-
-tmp_cotar_bceuropeu.py
-
-# C extensions
 *.so
-
-# Diretória Configurações do PyCharm
-.idea/
-.appbottle
-coverage/
-# Distribution / packaging
 .Python
-env/
+.secrets.toml
 build/
 develop-eggs/
 dist/
@@ -27,70 +15,63 @@ lib64/
 parts/
 sdist/
 var/
+wheels/
+share/python-wheels/
 *.egg-info/
 .installed.cfg
 *.egg
-
-# PyInstaller
-#  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+MANIFEST
 *.manifest
 *.spec
-
-# Installer logs
 pip-log.txt
 pip-delete-this-directory.txt
-
-# Unit test / coverage reports
 htmlcov/
 .tox/
+.nox/
 .coverage
 .coverage.*
-coverage.xml
 .cache
 nosetests.xml
-*,cover
+coverage.xml
+*.cover
+*.py,cover
 .hypothesis/
-
-# Translations
+.pytest_cache/
+cover/
 *.mo
 *.pot
-
-# Django stuff:
 *.log
 local_settings.py
-
-# Flask stuff:
+db.sqlite3
+db.sqlite3-journal
 instance/
 .webassets-cache
-
-# Scrapy stuff:
 .scrapy
-
-# Sphinx documentation
 docs/_build/
-
-# PyBuilder
+.pybuilder/
 target/
-
-# IPython Notebook
 .ipynb_checkpoints
-
-# pyenv
-.python-version
-
-# celery beat schedule file
+profile_default/
+ipython_config.py
+.pdm.toml
+__pypackages__/
 celerybeat-schedule
-
-# dotenv
+celerybeat.pid
+*.sage.py
 .env
-
-# virtualenv
+.venv
+env/
 venv/
 ENV/
-
-# Spyder project settings
+env.bak/
+venv.bak/
 .spyderproject
-
-# Rope project settings
+.spyproject
 .ropeproject
+/site
+.mypy_cache/
+.dmypy.json
+dmypy.json
+.pyre/
+.pytype/
+cython_debug/
