--- .travis.yml
@@ -1,15 +1,0 @@
-language: python
-python:
-  - 3.6.4
-install:
-  - pip install pipenv
-script:
-  - pipenv install --dev
-  - flake8 .
-  - pytest --cov=core
-after_success:
-  - codecov
-#notifications:
-#  slack:
-#    secure:
-#
