--- Procfile
@@ -1,1 +1,0 @@
-web: python core/app.py
