--- app.json
@@ -1,23 +1,0 @@
-{
-  "name": "cotacoes",
-  "scripts": {
-  },
-  "env": {
-    "APP_LOCATION": {
-      "required": true
-    },
-    "TZ": {
-      "required": true
-    }
-  },
-  "formation": {
-  },
-  "addons": [
-
-  ],
-  "buildpacks": [
-    {
-      "url": "heroku/python"
-    }
-  ]
-}
