+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "vmcotacoes"
-version = "0.1.1"
+version = "0.2.0"
 description = "Projeto de Cotações de Moedas Estrangeiras"
 authors = ["Vicente Marçal <vicente.marcal@gmail.com>"]
 license = "MIT"
@@ -8,17 +8,13 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
-Flask = "^2.2.2"
-requests = "^2.28.1"
 dynaconf = "^3.1.11"
-pytz = "^2022.4"
 
 
 [tool.poetry.group.dev.dependencies]
 codecov = "^2.1.12"
 flake8 = "^5.0.4"
 pytest-cov = "^4.0.0"
-pytest-flask = "^1.2.0"
 
 [tool.pytest.ini_options]
 python_files = [
