+++ rates/config.py
