+++ rates/frontend/static/css/stylesheet.css
