+++ rates/frontend/static/img/favicon.ico
