+++ rates/frontend/templates/base.html
