+++ rates/frontend/templates/error.html
