+++ rates/frontend/templates/index.html
