--- runtime
@@ -1,1 +1,0 @@
-python-3.6.4
